+++ src/factories/GameParser.h
@@ -16,7 +16,7 @@
 #include "../Exception.h"
 
 // amount, space, file, order, moveAmount, speed
-typedef std::tuple<int, int, std::string, std::string, int, int> infoTuple;
+typedef std::tuple<int, int, std::string, std::string, int, double> infoTuple;
 
 namespace factories {
 
@@ -42,7 +42,7 @@ private:
 
 	double speedMultiplier_;
 
-	void parseRow(TiXmlElement*, int, int);
+	void parseRow(TiXmlElement*, int, double);
 
 };
